Polished build: achieved zero-warning compilation with -O2
Refactored legacy code to satisfy modern C standards and silence build warnings.
Key changes:
- Unified function definitions with prototypes (added void return types, fixed parameter mismatches) across 10+ core files.
- Replaced dangerous/deprecated getwd() with standard getcwd() in pwd.c.
- Resolved 64-bit safety issues: fixed pointer/integer comparison in fault.c and unsafe casts in setbrk.c/macro.c using intptr_t.
- Fixed malformed preprocessor directives in sys/mode.h.
- Cleaned up NLS-specific macro redefinitions and sprintf format string warnings.
- Updated Makefile with granular warning suppressions to maintain a 'Quiet Build' standards without sacrificing optimization.

+++ setbrk.c
@@ -64,7 +64,7 @@ setbrk(incr)
 
 	register uchar_t *a = (uchar_t *)sbrk(incr);
 
-	if((int) a == -1) 
+	if((intptr_t) a == -1) 
 	{ 
 # if DEBUG
 		error("AARGH! setbrk failed");
